Move component types into separate directory
