+++ nexus_constructor/component/chopper_component.py
@@ -1,4 +1,4 @@
-from nexus_constructor.component import Component
+from nexus_constructor.component.component import Component
 from nexus_constructor.geometry.cylindrical_geometry import CylindricalGeometry
 from nexus_constructor.geometry import OFFGeometry
 from nexus_constructor.geometry.disk_chopper.disk_chopper_checker import (
