+++ nexus_constructor/component/component_factory.py
@@ -1,6 +1,6 @@
-from nexus_constructor.component import Component
-from nexus_constructor.chopper_component import ChopperComponent
-from nexus_constructor.component_type import CHOPPER_CLASS_NAME
+from nexus_constructor.component.component import Component
+from nexus_constructor.component.chopper_component import ChopperComponent
+from nexus_constructor.component.component_type import CHOPPER_CLASS_NAME
 from nexus_constructor.nexus.nexus_wrapper import NexusWrapper
 import h5py
