+++ nexus_constructor/component_tree_model.py
@@ -1,7 +1,7 @@
 from PySide2.QtCore import QAbstractItemModel, QModelIndex, Qt
 import PySide2.QtGui
 from PySide2.QtGui import QVector3D
-from nexus_constructor.component import Component
+from nexus_constructor.component.component import Component
 from nexus_constructor.transformations import Transformation, TransformationsList
 from nexus_constructor.instrument import Instrument
 from nexus_constructor.ui_utils import generate_unique_name
