+++ tests/test_component.py
@@ -1,6 +1,6 @@
 from mock import patch
 
-from nexus_constructor.component import (
+from nexus_constructor.component.component import (
     DependencyError,
     Component,
     SHAPE_GROUP_NAME,
@@ -423,7 +423,7 @@ def test_GIVEN_pixel_mapping_WHEN_setting_cylinder_shape_THEN_cylindrical_geomet
     pixel_mapping = PixelMapping()
 
     with patch(
-        "nexus_constructor.component.CylindricalGeometry"
+        "nexus_constructor.component.component.CylindricalGeometry"
     ) as mock_cylindrical_geometry_constructor:
         component.set_cylinder_shape(pixel_data=pixel_mapping)
         mock_cylindrical_geometry_constructor.assert_called_once_with(
@@ -440,7 +440,7 @@ def test_GIVEN_pixel_mapping_WHEN_setting_off_geometry_shape_THEN_off_geometry_i
     filename = "somefile.off"
 
     with patch(
-        "nexus_constructor.component.OFFGeometryNexus"
+        "nexus_constructor.component.component.OFFGeometryNexus"
     ) as mock_off_geometry_constructor:
 
         component.set_off_shape(
@@ -462,7 +462,7 @@ def test_GIVEN_no_pixel_data_WHEN_setting_cylinder_shape_THEN_shape_group_has_na
     component
 ):
     with patch(
-        "nexus_constructor.component.CylindricalGeometry"
+        "nexus_constructor.component.component.CylindricalGeometry"
     ) as mock_cylindrical_geometry_constructor:
         component.set_cylinder_shape(pixel_data=None)
         mock_cylindrical_geometry_constructor.assert_called_once_with(
@@ -478,7 +478,7 @@ def test_GIVEN_no_pixel_data_WHEN_setting_off_geometry_shape_THEN_shape_group_ha
     filename = "somefile.off"
 
     with patch(
-        "nexus_constructor.component.OFFGeometryNexus"
+        "nexus_constructor.component.component.OFFGeometryNexus"
     ) as mock_off_geometry_constructor:
 
         component.set_off_shape(
@@ -498,7 +498,7 @@ def test_GIVEN_pixel_grid_WHEN_setting_cylinder_shape_THEN_cylindrical_geometry_
     pixel_grid = PixelGrid()
 
     with patch(
-        "nexus_constructor.component.CylindricalGeometry"
+        "nexus_constructor.component.component.CylindricalGeometry"
     ) as mock_cylindrical_geometry_constructor:
         component.set_cylinder_shape(pixel_data=pixel_grid)
         mock_cylindrical_geometry_constructor.assert_called_once_with(
@@ -515,7 +515,7 @@ def test_GIVEN_pixel_grid_WHEN_setting_off_geometry_shape_THEN_off_geometry_is_n
     filename = "somefile.off"
 
     with patch(
-        "nexus_constructor.component.OFFGeometryNexus"
+        "nexus_constructor.component.component.OFFGeometryNexus"
     ) as mock_off_geometry_constructor:
 
         component.set_off_shape(
@@ -548,7 +548,7 @@ def test_GIVEN_off_properties_WHEN_setting_off_geometry_shape_THEN_shape_group_h
 ):
     off_geometry = OFFGeometryNoNexus(vertices=[], faces=[])
 
-    with patch("nexus_constructor.component.OFFGeometryNexus"):
+    with patch("nexus_constructor.component.component.OFFGeometryNexus"):
         component.set_off_shape(loaded_geometry=off_geometry)
 
     assert (
