+++ tests/test_component_fields.py
@@ -1,6 +1,6 @@
 from nexus_constructor.component_fields import add_fields_to_component
 from nexus_constructor.nexus.nexus_wrapper import NexusWrapper
-from nexus_constructor.component import Component
+from nexus_constructor.component.component import Component
 import numpy as np
