+++ tests/ui_tests/test_ui_add_component_window.py
@@ -11,9 +11,9 @@ from mock import Mock, call, patch, mock_open
 from pytestqt.qtbot import QtBot
 import numpy as np
 
-from nexus_constructor import component_type
+from nexus_constructor.component import component_type
 from nexus_constructor.add_component_window import AddComponentDialog
-from nexus_constructor.component import Component
+from nexus_constructor.component.component import Component
 from nexus_constructor.component_tree_model import ComponentTreeModel
 from nexus_constructor.geometry import OFFGeometryNoNexus
 from nexus_constructor.instrument import Instrument
